LPD-28412 LPD-32914 portal-search-test: refactor into Base Test
https://liferay.atlassian.net/browse/LPD-28412
